Added project id as variable, and setup full deployment flow without errors  (#114)

+++ services/booking/pkg/models.go
@@ -34,10 +34,10 @@ type WorkingHours struct {
 }
 
 type PaymentRequest struct {
-	UserID             int64   `json:"user_id"`
-	UserEmail          string  `json:"user_email"`
-	IdempotencyKey     string  `json:"idempotency_key"`
-	Amount             float32 `json:"amount"`
-	PaymentDescription string  `json:"payment_description"`
-	PaymentType        string  `json:"payment_type"`
+	UserID             int64  `json:"user_id"`
+	UserEmail          string `json:"user_email"`
+	IdempotencyKey     string `json:"idempotency_key"`
+	Amount             int64  `json:"amount"`
+	PaymentDescription string `json:"payment_description"`
+	PaymentType        string `json:"payment_type"`
 }
